+++ test/index.php
@@ -1,0 +1,12 @@
+<?php
+
+/**
+ * All unit test right here
+ */
+
+// Unit Test To Run
+
+// Image Uploader Test
+require_once("./uploader/test.php");
+
+// Add Other Test